Get rid of <Paginator /> component
