+++ saleor/static/dashboard-next/collections/views/CollectionDetails.tsx
@@ -4,13 +4,12 @@ import * as React from "react";
 
 import ActionDialog from "../../components/ActionDialog";
 import AssignProductDialog from "../../components/AssignProductDialog";
-import { createPaginationState } from "../../components/Paginator";
 import { WindowTitle } from "../../components/WindowTitle";
 import { SearchProductsProvider } from "../../containers/SearchProducts";
 import useBulkActions from "../../hooks/useBulkActions";
 import useNavigator from "../../hooks/useNavigator";
 import useNotifier from "../../hooks/useNotifier";
-import usePaginator from "../../hooks/usePaginator";
+import usePaginator, { createPaginationState } from "../../hooks/usePaginator";
 import i18n from "../../i18n";
 import { getMutationState, maybe } from "../../misc";
 import { productUrl } from "../../products/urls";
