+++ saleor/static/dashboard-next/discounts/views/SaleDetails.tsx
@@ -8,7 +8,6 @@ import ActionDialog from "../../components/ActionDialog";
 import AssignCategoriesDialog from "../../components/AssignCategoryDialog";
 import AssignCollectionDialog from "../../components/AssignCollectionDialog";
 import AssignProductDialog from "../../components/AssignProductDialog";
-import { createPaginationState } from "../../components/Paginator";
 import { WindowTitle } from "../../components/WindowTitle";
 import { SearchCategoriesProvider } from "../../containers/SearchCategories";
 import { SearchCollectionsProvider } from "../../containers/SearchCollections";
@@ -16,7 +15,7 @@ import { SearchProductsProvider } from "../../containers/SearchProducts";
 import useBulkActions from "../../hooks/useBulkActions";
 import useNavigator from "../../hooks/useNavigator";
 import useNotifier from "../../hooks/useNotifier";
-import usePaginator from "../../hooks/usePaginator";
+import usePaginator, { createPaginationState } from "../../hooks/usePaginator";
 import useShop from "../../hooks/useShop";
 import i18n from "../../i18n";
 import { decimal, getMutationState, maybe } from "../../misc";
