+++ saleor/static/dashboard-next/orders/views/OrderList.tsx
@@ -1,11 +1,10 @@
 import Button from "@material-ui/core/Button";
 import * as React from "react";
 
-import { createPaginationState } from "../../components/Paginator";
 import useBulkActions from "../../hooks/useBulkActions";
 import useNavigator from "../../hooks/useNavigator";
 import useNotifier from "../../hooks/useNotifier";
-import usePaginator from "../../hooks/usePaginator";
+import usePaginator, { createPaginationState } from "../../hooks/usePaginator";
 import i18n from "../../i18n";
 import { getMutationState, maybe } from "../../misc";
 import { OrderStatusFilter } from "../../types/globalTypes";
